+++ ts/components/game/Hero.tsx
@@ -1,16 +1,16 @@
 import * as React from "react";
 import * as Immutable from "immutable";
 import EntityInPlay from "./EntityInPlay";
-import {EntityInPlayProps, CardOracleProps} from "../../interfaces";
+import {CardOracleProps, EntityInPlayProps} from "../../interfaces";
 import Entity from "../../Entity";
 import Attack from "./stats/Attack";
 import Damage from "./stats/Damage";
 import Healing from "./stats/Healing";
 import Health from "./stats/Health";
 import Armor from "./stats/Armor";
-import SecretText from "./stats/SecretText"
+import SecretText from "./stats/SecretText";
 import HeroArt from "./visuals/HeroArt";
-import {MetaDataType} from "../../enums";
+import {GameTag, MetaDataType} from "../../enums";
 import MetaData from "../../MetaData";
 import GameStateDescriptor from "../../state/GameStateDescriptor";
 
@@ -26,22 +26,30 @@ export default class Hero extends EntityInPlay<HeroProps> {
 	protected jsx() {
 		let entity = this.props.entity;
 
-		let secretCount = this.props.secrets.count();
-		let secretText = secretCount > 1 ? secretCount.toString() : "?";
+		let secrets = this.props.secrets;
+		let quests = [];
+
+		let hasQuest = secrets.some((potentialQuest: Entity) => !!potentialQuest.getTag(GameTag.QUEST));
+		let secretCount = secrets.count();
+
+		// build text in icon
+		let secretText = hasQuest ? "!" : (secretCount > 1 ? "" + secretCount : "?");
+
+		// build title
 		let secretTitle = this.props.secrets.reduce((title, entity: Entity): string => {
 			let name = entity.cardId;
-			if(!entity.revealed) {
-				if(this.props.cardOracle && this.props.cardOracle.has(entity.id)) {
+			if (!entity.revealed) {
+				if (this.props.cardOracle && this.props.cardOracle.has(entity.id)) {
 					name = this.props.cardOracle.get(entity.id);
 				}
 				else {
 					return title;
 				}
 			}
-			if(title) {
+			if (title) {
 				title += ", ";
 			}
-			if(this.props.cards && this.props.cards.has(name)) {
+			if (this.props.cards && this.props.cards.has(name)) {
 				name = this.props.cards.get(name).name || name;
 			}
 			return title += name;
@@ -54,7 +62,7 @@ export default class Hero extends EntityInPlay<HeroProps> {
 			this.props.descriptors.forEach((descriptor: GameStateDescriptor) => {
 				descriptor.metaData.forEach((metaData: MetaData) => {
 					if (metaData.entities.has(entity.id)) {
-						switch(metaData.type) {
+						switch (metaData.type) {
 							case MetaDataType.DAMAGE:
 								damage += metaData.data;
 								break;
@@ -68,7 +76,8 @@ export default class Hero extends EntityInPlay<HeroProps> {
 		}
 
 		return [
-			<HeroArt key="art"
+			<HeroArt
+				key="art"
 				entity={entity}
 				secrets={this.props.secrets}
 				cards={this.props.cards}
@@ -76,15 +85,15 @@ export default class Hero extends EntityInPlay<HeroProps> {
 				cardArtDirectory={this.props.cardArtDirectory}
 				damage={damage}
 				healing={healing}
-				/>,
+			/>,
 			<div key="stats" className="stats">
-				{entity.getAtk() ? <Attack attack={entity.getAtk()}/> : null}
-				<Health health={entity.getHealth() } damage={entity.getDamage()}/>
-				{entity.getArmor() ? <Armor armor={entity.getArmor()}/> : null}
-				{secretCount > 0 ? <SecretText text={secretText} title={secretTitle}/> : null}
-				{damage != 0 ? <Damage damage={damage}/> : null}
-				{healing != 0 ? <Healing healing={healing}/> : null}
-			</div>
+				{entity.getAtk() ? <Attack attack={entity.getAtk()} /> : null}
+				<Health health={entity.getHealth() } damage={entity.getDamage()} />
+				{entity.getArmor() ? <Armor armor={entity.getArmor()} /> : null}
+				{(hasQuest || secretCount > 0) ? <SecretText text={secretText} title={secretTitle} /> : null}
+				{damage != 0 ? <Damage damage={damage} /> : null}
+				{healing != 0 ? <Healing healing={healing} /> : null}
+			</div>,
 		];
 	}
 }